requirements.txt
streamlit
openai
python-docx
python-dotenv
PyPDF2